software-mandala/MANDALA APP/backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.1
dj-database-url==2.2.0
Django==5.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.8.0
django-filter==25.2
djangorestframework==3.16.1
gunicorn==22.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
python-slugify==8.0.4
pytz==2025.2
redis==6.4.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
whitenoise==6.11.0
django-axes==6.5.2
drf-spectacular==0.28.0
